extract export entries function into separate file
